Linear filter model, respecting cell classes
This is as far as I can get with the restriction that the filtering happens independently to the midget and parasol classes. It is an informative failure: there does not seem to be a way to combine this information to recover the cortical TTFs. This would imply that there is filtering that happens at the cortex AFTER the combination of midget and parasol luminance signals.